requirements.txt
streamlit
faiss-cpu
sentence-transformers
cohere
python-dotenv
PyPDF2
